backend/requirements.txt
Django
djangorestframework
django-cors-headers
djangorestframework-simplejwt
python-dotenv
